fix integration test not working
