+++ buildtools/ci.buildspec.yml
@@ -6,7 +6,7 @@ phases:
       dotnet: 8.x
   build:
     commands:
-      - dotnet test --verbosity normal integrations-on-dotnet-aspire-for-aws.sln --configuration Release --logger trx --results-directory ./testresults
+      - dotnet test --verbosity normal integrations-on-dotnet-aspire-for-aws.sln --configuration Release --logger "console;verbosity=detailed" --logger trx --results-directory ./testresults
 reports:
     integrations-on-dotnet-aspire-for-aws-tests:
         file-format: VisualStudioTrx
