+++ src/Aspire.Hosting.AWS/Lambda/LambdaExtensions.cs
@@ -10,7 +10,6 @@ using Microsoft.Extensions.DependencyInjection.Extensions;
 using System.Diagnostics;
 using System.Net.Sockets;
 using System.Runtime.Versioning;
-using System.Text;
 using Aspire.Hosting.AWS.Utils;
 
 #pragma warning disable IDE0130
@@ -61,42 +60,12 @@ public static class LambdaExtensions
             {
                 var project = new LambdaProjectResource(name);
                 resource = builder.AddResource(project)
-                    // .WithAnnotation(new LambdaProjectMetadata(projectPath))
                     .WithAnnotation(new TLambdaProject());
                 var projectMetadata = project.Annotations.OfType<IProjectMetadata>().First();
                 project.Annotations.Remove(projectMetadata);
-                string tempPath = Path.Combine(Path.GetTempPath(), Guid.NewGuid().ToString());
-                Directory.CreateDirectory(tempPath);
 
-                // Define the .csproj content.
-                var projectContent = $@"
-<Project Sdk=""Microsoft.NET.Sdk"">
-  <PropertyGroup>
-    <OutputType>Exe</OutputType>
-    <TargetFramework>net8.0</TargetFramework>
-  </PropertyGroup>
-  <ItemGroup>
-    <PackageReference Include=""Amazon.Lambda.RuntimeSupport"" Version=""1.12.2"" />
-  </ItemGroup>
-  <ItemGroup>
-    <ProjectReference Include=""{projectMetadata.ProjectPath}"" />
-  </ItemGroup>
-</Project>
-";
-                var projectName = $"Wrapper{Path.GetFileName(projectMetadata.ProjectPath)}";
-                var projectPath = Path.Combine(tempPath, projectName);
-                File.WriteAllText(projectPath, projectContent);
-                
-                string programContent = $@"
-using Amazon.Lambda.RuntimeSupport;
-
-RuntimeSupportInitializer runtimeSupportInitializer = new RuntimeSupportInitializer(""{lambdaHandler}"");
-await runtimeSupportInitializer.RunLambdaBootstrap();
-";
-                var programPath = Path.Combine(tempPath, "Program.cs");
-                File.WriteAllText(programPath, programContent);
-
-                RunProcess("dotnet", $"build {new FileInfo(projectPath).Name}", Directory.GetParent(projectPath)!.FullName);
+                var projectPath =
+                    ProjectUtilities.CreateExecutableWrapperProject(projectMetadata.ProjectPath, lambdaHandler);
                 
                 resource = resource
                     .WithAnnotation(new LambdaProjectMetadata(projectPath));
@@ -250,56 +219,5 @@ await runtimeSupportInitializer.RunLambdaBootstrap();
         builder.WithOtlpExporter();
 
         return builder;
-    }
-    
-    /// <summary>
-    /// Utility method for running a command on the commandline. It returns backs the exit code and anything written to stdout or stderr.
-    /// </summary>
-    /// <param name="logger"></param>
-    /// <param name="path"></param>
-    /// <param name="arguments"></param>
-    /// <param name="cancellationToken"></param>
-    /// <returns></returns>
-    public static int RunProcess(string path, string arguments, string workingDirectory)
-    {
-        using var process = new Process
-        {
-            StartInfo = new ProcessStartInfo
-            {
-                WorkingDirectory = workingDirectory,
-                FileName = path,
-                Arguments = arguments,
-                RedirectStandardOutput = true,
-                RedirectStandardError = true,
-                UseShellExecute = false,
-                WindowStyle = ProcessWindowStyle.Hidden
-            }
-        };
-
-        var output = new StringBuilder();
-
-        process.OutputDataReceived += (sender, e) =>
-        {
-            if (e.Data != null)
-            {
-                output.Append(e.Data);
-            }
-        };
-
-        process.ErrorDataReceived += (sender, e) =>
-        {
-            if (e.Data != null)
-            {
-                output.Append(e.Data);
-            }
-        };
-
-        process.Start();
-        process.BeginOutputReadLine();
-        process.BeginErrorReadLine();
-
-        process.WaitForExit(int.MaxValue);
-
-        return process.ExitCode;
     }
 }
